+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Notable `mc-server-wrapper` changes, tracked in the [keep a changelog](https://k
 ### Internal
 
 * Updated to `twilight` 0.1 release from crates.io
+* Replaced custom mention parsing code with the new `twilight-mention` support for parsing mentions from Discord messages
 
 ## [alpha2] - 2020-08-22